% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncovaRobustWrapper.R, R/ancovaWrapper.R
\name{AncovaRobust}
\alias{AncovaRobust}
\title{Robust ANCOVA}
\usage{
AncovaRobust(
  data = NULL,
  version = "0.19",
  formula = NULL,
  contrastCi = FALSE,
  contrastCiLevel = 0.95,
  contrasts = list(optionKey = "variable", types = list(), value = list()),
  covariates = list(types = list(), value = list()),
  dependent = list(types = list(), value = ""),
  descriptivePlotCiLevel = 0.95,
  descriptivePlotErrorBar = FALSE,
  descriptivePlotErrorBarType = "ci",
  descriptivePlotHorizontalAxis = list(types = list(), value = ""),
  descriptivePlotSeparateLines = list(types = list(), value = ""),
  descriptivePlotSeparatePlot = list(types = list(), value = ""),
  fixedFactors = list(types = list(), value = list()),
  modelTerms = list(optionKey = "components", types = list(), value = list()),
  postHocCorrectionBonferroni = FALSE,
  postHocCorrectionHolm = FALSE,
  postHocCorrectionScheffe = FALSE,
  postHocCorrectionTukey = TRUE,
  postHocTerms = list(optionKey = "variable", types = list(), value = list()),
  rainCloudHorizontalAxis = list(types = list(), value = ""),
  rainCloudHorizontalDisplay = FALSE,
  rainCloudSeparatePlots = list(types = list(), value = ""),
  rainCloudYAxisLabel = "",
  wlsWeights = list(types = list(), value = "")
)

AncovaRobust(
  data = NULL,
  version = "0.19",
  formula = NULL,
  contrastCi = FALSE,
  contrastCiLevel = 0.95,
  contrasts = list(optionKey = "variable", types = list(), value = list()),
  covariates = list(types = list(), value = list()),
  dependent = list(types = list(), value = ""),
  descriptivePlotCiLevel = 0.95,
  descriptivePlotErrorBar = FALSE,
  descriptivePlotErrorBarType = "ci",
  descriptivePlotHorizontalAxis = list(types = list(), value = ""),
  descriptivePlotSeparateLines = list(types = list(), value = ""),
  descriptivePlotSeparatePlot = list(types = list(), value = ""),
  fixedFactors = list(types = list(), value = list()),
  modelTerms = list(optionKey = "components", types = list(), value = list()),
  postHocCorrectionBonferroni = FALSE,
  postHocCorrectionHolm = FALSE,
  postHocCorrectionScheffe = FALSE,
  postHocCorrectionTukey = TRUE,
  postHocTerms = list(optionKey = "variable", types = list(), value = list()),
  rainCloudHorizontalAxis = list(types = list(), value = ""),
  rainCloudHorizontalDisplay = FALSE,
  rainCloudSeparatePlots = list(types = list(), value = ""),
  rainCloudYAxisLabel = "",
  wlsWeights = list(types = list(), value = "")
)
}
\arguments{
\item{covariates, }{An optional continuous covariate to adjust for.}

\item{dependent, }{The continuous outcome variable.}

\item{descriptivePlotErrorBar, }{Display error bars in the plot. The error bars can represent either confidence intervals or standard errors.
Defaults to \code{TRUE}.}

\item{descriptivePlotHorizontalAxis, }{Select the independent variable that should be displayed on the horizontal axis of the plot.}

\item{descriptivePlotSeparateLines, }{By placing an independent variable in this box, different lines corresponding to the different levels of the independent variable will be displayed.}

\item{descriptivePlotSeparatePlot, }{By placing an independent variable in this box, different plots corresponding to the different levels of the independent variable will be displayed.}

\item{fixedFactors, }{A categorical grouping variable with exactly two levels.}

\item{rainCloudHorizontalAxis, }{Select the independent variable that should be displayed on the horizontal axis of the plot.}

\item{rainCloudHorizontalDisplay, }{Changes the orientation of the raincloud difference plot so that the x-axis represents the dependent variable.
Defaults to \code{FALSE}.}

\item{rainCloudSeparatePlots, }{By placing an independent variable in this box, different plots corresponding to the different levels of the independent variable will be displayed.}

\item{bootstrapSamples, }{The number of bootstrap resamples to use. Default is 599.}

\item{descriptivesTable, }{Per-group sample size, trimmed mean, median, Winsorized standard deviation, and median absolute deviation for the dependent variable and each covariate.
Defaults to \code{FALSE}.}

\item{robustMethod, }{Select the robust estimation method to use for the analysis.}

\item{trimProportion, }{The proportion of observations to trim from each tail. Default is 0.2 (20%% trimming).}
}
\description{
Robust ANCOVA compares trimmed means of exactly two groups while adjusting for a covariate, using a nonparametric running interval smoother. No parametric assumptions about homogeneity of regression slopes or equal variances are required. Without a covariate, a robust one-way ANOVA is performed instead.
## Assumptions
- The independent variable is categorical with exactly two levels.
- The dependent variable and covariate are continuous.
- The groups are independent.
}
