% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnovaRepeatedMeasuresRobustWrapper.R,
%   R/anovarepeatedmeasuresWrapper.R
\name{AnovaRepeatedMeasuresRobust}
\alias{AnovaRepeatedMeasuresRobust}
\title{Robust Repeated Measures ANOVA}
\usage{
AnovaRepeatedMeasuresRobust(
  data = NULL,
  version = "0.19",
  formula = NULL,
  betweenModelTerms = list(optionKey = "components", types = list(), value = list()),
  betweenSubjectFactors = list(types = list(), value = list()),
  contrastCi = FALSE,
  contrastCiLevel = 0.95,
  contrasts = list(optionKey = "variable", types = list(), value = list()),
  covariates = list(types = list(), value = list()),
  descriptivePlotCiLevel = 0.95,
  descriptivePlotErrorBar = FALSE,
  descriptivePlotErrorBarType = "ci",
  descriptivePlotHorizontalAxis = list(types = list(), value = ""),
  descriptivePlotSeparateLines = list(types = list(), value = ""),
  descriptivePlotSeparatePlot = list(types = list(), value = ""),
  postHocCorrectionBonferroni = FALSE,
  postHocCorrectionHolm = FALSE,
  postHocCorrectionScheffe = FALSE,
  postHocCorrectionTukey = TRUE,
  postHocTerms = list(optionKey = "variable", types = list(), value = list()),
  rainCloudHorizontalAxis = list(types = list(), value = ""),
  rainCloudHorizontalDisplay = FALSE,
  rainCloudSeparatePlots = list(types = list(), value = ""),
  rainCloudYAxisLabel = "",
  repeatedMeasuresCells = list(types = list(), value = list()),
  repeatedMeasuresFactors = list(),
  withinModelTerms = list(optionKey = "components", types = list(), value = list())
)

AnovaRepeatedMeasuresRobust(
  data = NULL,
  version = "0.19",
  formula = NULL,
  betweenModelTerms = list(optionKey = "components", types = list(), value = list()),
  betweenSubjectFactors = list(types = list(), value = list()),
  contrastCi = FALSE,
  contrastCiLevel = 0.95,
  contrasts = list(optionKey = "variable", types = list(), value = list()),
  covariates = list(types = list(), value = list()),
  descriptivePlotCiLevel = 0.95,
  descriptivePlotErrorBar = FALSE,
  descriptivePlotErrorBarType = "ci",
  descriptivePlotHorizontalAxis = list(types = list(), value = ""),
  descriptivePlotSeparateLines = list(types = list(), value = ""),
  descriptivePlotSeparatePlot = list(types = list(), value = ""),
  postHocCorrectionBonferroni = FALSE,
  postHocCorrectionHolm = FALSE,
  postHocCorrectionScheffe = FALSE,
  postHocCorrectionTukey = TRUE,
  postHocTerms = list(optionKey = "variable", types = list(), value = list()),
  rainCloudHorizontalAxis = list(types = list(), value = ""),
  rainCloudHorizontalDisplay = FALSE,
  rainCloudSeparatePlots = list(types = list(), value = ""),
  rainCloudYAxisLabel = "",
  repeatedMeasuresCells = list(types = list(), value = list()),
  repeatedMeasuresFactors = list(),
  withinModelTerms = list(optionKey = "components", types = list(), value = list())
)
}
\arguments{
\item{betweenSubjectFactors, }{Optional single between-subjects grouping variable for mixed designs.}

\item{descriptivePlotErrorBar, }{Display error bars in the plot. The error bars can represent either confidence intervals or standard errors.
Defaults to \code{TRUE}.}

\item{descriptivePlotHorizontalAxis, }{Select the independent variable that should be displayed on the horizontal axis of the plot.}

\item{descriptivePlotSeparateLines, }{By placing an independent variable in this box, different lines corresponding to the different levels of the independent variable will be displayed.}

\item{descriptivePlotSeparatePlot, }{By placing an independent variable in this box, different plots corresponding to the different levels of the independent variable will be displayed.}

\item{postHocCorrectionBonferroni, }{This correction is considered conservative. The risk of Type I error is reduced, however the statistical power decreases as well.
Defaults to \code{FALSE}.}

\item{postHocCorrectionHolm, }{Also called sequential Bonferroni, and considered less conservative than the Bonferroni method.
Defaults to \code{FALSE}.}

\item{rainCloudHorizontalAxis, }{Select the independent variable that should be displayed on the horizontal axis of the plot.}

\item{rainCloudHorizontalDisplay, }{Changes the orientation of the raincloud difference plot so that the x-axis represents the dependent variable.
Defaults to \code{FALSE}.}

\item{rainCloudSeparatePlots, }{By placing an independent variable in this box, different plots corresponding to the different levels of the independent variable will be displayed.}

\item{repeatedMeasuresCells, }{The separate columns in the data frame that represent the levels of the repeated measure(s) factor(s).}

\item{repeatedMeasuresFactors, }{The within-subjects (repeated measures) variable. Define one factor and label its levels. Robust RM ANOVA currently supports a single within-subjects factor.}

\item{bootstrapSamples, }{The number of bootstrap resamples to use. Default is 599.}

\item{descriptivesTable, }{Per-cell sample size, trimmed mean, median, Winsorized standard deviation, and median absolute deviation across the within-subjects (and optional between-subjects) factor levels.
Defaults to \code{FALSE}.}

\item{postHocCi, }{When this option is selected, the confidence interval for the trimmed mean difference is displayed.
Defaults to \code{FALSE}.}

\item{postHocCorrectionHochberg, }{Hochberg's (1988) sharper Bonferroni procedure. This is the default correction used by WRS2's lincon function.
Defaults to \code{TRUE}.}

\item{postHocSignificanceFlag, }{Add asterisks to the table to indicate 3 levels of significance.
Defaults to \code{FALSE}.}

\item{robustMethod, }{Select the robust estimation method to use for the analysis.}

\item{trimProportion, }{The proportion of observations to trim from each tail. Default is 0.2 (20%% trimming).}
}
\description{
Robust Repeated Measures ANOVA analyzes differences between dependent group means using robust methods from the WRS2 package. No assumptions of normality, homogeneity, or sphericity are required.
## Assumptions
- The dependent variable is continuous.
- The observations within each subject are dependent (repeated measures).
}
