% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnovaRobustWrapper.R, R/anovaWrapper.R
\name{AnovaRobust}
\alias{AnovaRobust}
\title{Robust ANOVA}
\usage{
AnovaRobust(
  data = NULL,
  version = "0.19",
  formula = NULL,
  contrastCi = FALSE,
  contrastCiLevel = 0.95,
  contrasts = list(optionKey = "variable", types = list(), value = list()),
  dependent = list(types = list(), value = ""),
  descriptivePlotCiLevel = 0.95,
  descriptivePlotErrorBar = FALSE,
  descriptivePlotErrorBarType = "ci",
  descriptivePlotHorizontalAxis = list(types = list(), value = ""),
  descriptivePlotSeparateLines = list(types = list(), value = ""),
  descriptivePlotSeparatePlot = list(types = list(), value = ""),
  fixedFactors = list(types = list(), value = list()),
  modelTerms = list(optionKey = "components", types = list(), value = list()),
  postHocCorrectionBonferroni = FALSE,
  postHocCorrectionHolm = FALSE,
  postHocCorrectionScheffe = FALSE,
  postHocCorrectionTukey = TRUE,
  postHocTerms = list(optionKey = "variable", types = list(), value = list()),
  rainCloudHorizontalAxis = list(types = list(), value = ""),
  rainCloudHorizontalDisplay = FALSE,
  rainCloudSeparatePlots = list(types = list(), value = ""),
  rainCloudYAxisLabel = "",
  wlsWeights = list(types = list(), value = "")
)

AnovaRobust(
  data = NULL,
  version = "0.19",
  formula = NULL,
  contrastCi = FALSE,
  contrastCiLevel = 0.95,
  contrasts = list(optionKey = "variable", types = list(), value = list()),
  dependent = list(types = list(), value = ""),
  descriptivePlotCiLevel = 0.95,
  descriptivePlotErrorBar = FALSE,
  descriptivePlotErrorBarType = "ci",
  descriptivePlotHorizontalAxis = list(types = list(), value = ""),
  descriptivePlotSeparateLines = list(types = list(), value = ""),
  descriptivePlotSeparatePlot = list(types = list(), value = ""),
  fixedFactors = list(types = list(), value = list()),
  modelTerms = list(optionKey = "components", types = list(), value = list()),
  postHocCorrectionBonferroni = FALSE,
  postHocCorrectionHolm = FALSE,
  postHocCorrectionScheffe = FALSE,
  postHocCorrectionTukey = TRUE,
  postHocTerms = list(optionKey = "variable", types = list(), value = list()),
  rainCloudHorizontalAxis = list(types = list(), value = ""),
  rainCloudHorizontalDisplay = FALSE,
  rainCloudSeparatePlots = list(types = list(), value = ""),
  rainCloudYAxisLabel = "",
  wlsWeights = list(types = list(), value = "")
)
}
\arguments{
\item{dependent, }{The continuous outcome variable.}

\item{descriptivePlotErrorBar, }{Display error bars in the plot. The error bars can represent either confidence intervals or standard errors.
Defaults to \code{TRUE}.}

\item{descriptivePlotHorizontalAxis, }{Select the independent variable that should be displayed on the horizontal axis of the plot.}

\item{descriptivePlotSeparateLines, }{By placing an independent variable in this box, different lines corresponding to the different levels of the independent variable will be displayed.}

\item{descriptivePlotSeparatePlot, }{By placing an independent variable in this box, different plots corresponding to the different levels of the independent variable will be displayed.}

\item{fixedFactors, }{The categorical grouping variables. Trimmed means support up to three factors; medians support up to two; the bootstrap method supports a single factor (one-way) only.}

\item{postHocCorrectionBonferroni, }{This correction is considered conservative. The risk of Type I error is reduced, however the statistical power decreases as well.
Defaults to \code{FALSE}.}

\item{postHocCorrectionHolm, }{Also called sequential Bonferroni, and considered less conservative than the Bonferroni method.
Defaults to \code{FALSE}.}

\item{rainCloudHorizontalAxis, }{Select the independent variable that should be displayed on the horizontal axis of the plot.}

\item{rainCloudHorizontalDisplay, }{Changes the orientation of the raincloud difference plot so that the x-axis represents the dependent variable.
Defaults to \code{FALSE}.}

\item{rainCloudSeparatePlots, }{By placing an independent variable in this box, different plots corresponding to the different levels of the independent variable will be displayed.}

\item{bootstrapSamples, }{The number of bootstrap resamples to use. Default is 599.}

\item{descriptivesTable, }{Per-cell sample size, trimmed (or untrimmed) mean, median, Winsorized standard deviation, and median absolute deviation.
Defaults to \code{FALSE}.}

\item{effectSizeTable, }{Algina-Keselman-Penfield robust standardised mean differences (ξ) for each pair of factor levels. Computed via WRS2::akp.effect on a one-way design.
Defaults to \code{FALSE}.}

\item{postHocCi, }{When this option is selected, the confidence interval for the trimmed mean difference is displayed.
Defaults to \code{FALSE}.}

\item{postHocCorrectionHochberg, }{Hochberg's (1988) sharper Bonferroni procedure. This is the default correction used by WRS2's lincon function.
Defaults to \code{TRUE}.}

\item{postHocSignificanceFlag, }{Add asterisks to the table to indicate 3 levels of significance.
Defaults to \code{FALSE}.}

\item{robustMethod, }{Select the robust estimation method to use for the analysis.}

\item{trimProportion, }{The proportion of observations to trim from each tail. Default is 0.2 (20%% trimming).}
}
\description{
Robust ANOVA analyzes differences between group means using robust methods from the WRS2 package. These methods do not assume normality or homogeneity of variances.
## Assumptions
- The independent variables are categorical, the dependent variable is continuous.
- The groups are independent.
}
